dialog refactor continues
