+++ pytc_gui/dialog/__init__.py
@@ -5,4 +5,6 @@ Submodule with dialog boxes raised by main window.
 __author__ = "Michael J. Harms"
 __date__ = "2017-06-01"
 
-from .aic_test import DoAICTest
+from .aic_test import AICTest
+from .documentation import Documentation
+from .about import About
